added amd - must be checked.

+++ src/sparse/utils/mod.rs
@@ -19,14 +19,14 @@ where
     let mut new_non_zero_count = non_zero_count;
 
     for p in matrix_a.col_ptr()[col_index]..matrix_a.col_ptr()[col_index + 1] {
-        row_index = matrix_a.row_idx()[p];
+        row_index = matrix_a.row_idx()[p as usize];
         if row_marker[row_index] < marker_value {
             row_marker[row_index] = marker_value;
             result_matrix.row_idx_mut()[new_non_zero_count] = row_index;
             new_non_zero_count += 1;
-            row_values[row_index] = scalar * matrix_a.values()[p];
+            row_values[row_index] = scalar * matrix_a.values()[p as usize];
         } else {
-            row_values[row_index] += scalar * matrix_a.values()[p];
+            row_values[row_index] += scalar * matrix_a.values()[p as usize];
         }
     }
